# Clut scan speedup