Processing time of the function which scans the clut folder structure when starting rt (as usual all measures are median of 7):

dev: 52.5 ms
1st commit: 24.0 ms
2nd commit: 23.0 ms
3rd commit: 22.6 ms