@heckflosse 2019-05-13
Edit: The 30 ms speedup when starting rt doesnt sound much, but when added to the speedup in the adjuster_review branch it's already a 150 ms speedup when starting rt in setm mode.

Small speedups sum up...